moves countdown from sessionstarted to rtiview
